+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.0.0-beta.14
 ### Breaking
 * remove `zip` function that just `(map list l1 l2 ...)`
+* `raise` now throws object as is (just like JS `throw` keyword)
+* `error` now throws LipsError exception as per R7RS
 ### Features
 * R7RS datum labels
 * allow to use `set-repr!` on records (record type name is now a class)
@@ -9,6 +11,7 @@
 * add codemirror hits for std lib symbols
 * experimental compiler to JSON (not 100% ready)
 * add support for `exp` on complex numbers
+* add R7RS `error-object?`, `error-object-message` and `error-object-irritants`
 ### Bugfix
 * fix `set-repr!` on classes
 * fix conflict with jQuery plugns in bookmark